Move the core of OpenRA.Mods.RA.Scripting to OpenRA.Mods.Common.Scripting
