+++ OpenRA.Mods.Common/CommonTraitsInterfaces.cs
@@ -21,6 +21,7 @@ namespace OpenRA.Mods.Common
 
 	public interface INotifyChat { bool OnChat(string from, string message); }
 	public interface IRenderActorPreviewInfo { IEnumerable<IActorPreview> RenderPreview (ActorPreviewInitializer init); }
+	public interface ICruiseAltitudeInfo { WRange GetCruiseAltitude(); }
 
 	public interface IUpgradable
 	{
