+++ OpenRA.Mods.Common/OpenRA.Mods.Common.csproj
@@ -89,6 +89,24 @@
     <Compile Include="ServerTraits\LobbySettingsNotification.cs" />
     <Compile Include="ServerTraits\MasterServerPinger.cs" />
     <Compile Include="ServerTraits\PlayerPinger.cs" />
+    <Compile Include="Scripting\CallLuaFunc.cs" />
+    <Compile Include="Scripting\LuaScript.cs" />
+    <Compile Include="Scripting\Media.cs" />
+    <Compile Include="Scripting\ScriptTriggers.cs" />
+    <Compile Include="Scripting\Global\ActorGlobal.cs" />
+    <Compile Include="Scripting\Global\CameraGlobal.cs" />
+    <Compile Include="Scripting\Global\CoordinateGlobals.cs" />
+    <Compile Include="Scripting\Global\DateTimeGlobal.cs" />
+    <Compile Include="Scripting\Global\MapGlobal.cs" />
+    <Compile Include="Scripting\Global\MediaGlobal.cs" />
+    <Compile Include="Scripting\Global\PlayerGlobal.cs" />
+    <Compile Include="Scripting\Global\TriggerGlobal.cs" />
+    <Compile Include="Scripting\Global\UtilsGlobal.cs" />
+    <Compile Include="Scripting\Properties\DiplomacyProperties.cs" />
+    <Compile Include="Scripting\Properties\HealthProperties.cs" />
+    <Compile Include="Scripting\Properties\PowerProperties.cs" />
+    <Compile Include="Scripting\Properties\ResourceProperties.cs" />
+    <Compile Include="Scripting\Properties\UpgradeProperties.cs" />
     <Compile Include="Traits\BlocksBullets.cs" />
     <Compile Include="Traits\Burns.cs" />
     <Compile Include="Traits\CustomBuildTimeValue.cs" />
