+++ OpenRA.Mods.Common/Scripting/Global/CameraGlobal.cs
@@ -10,7 +10,7 @@
 
 using OpenRA.Scripting;
 
-namespace OpenRA.Mods.RA.Scripting
+namespace OpenRA.Mods.Common.Scripting
 {
 	[ScriptGlobal("Camera")]
 	public class CameraGlobal : ScriptGlobal
