+++ OpenRA.Mods.RA/Widgets/Logic/Ingame/LeaveMapLogic.cs
@@ -11,8 +11,8 @@
 using System;
 using System.Drawing;
 using System.Linq;
+using OpenRA.Mods.Common.Scripting;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Scripting;
 using OpenRA.Network;
 using OpenRA.Traits;
 using OpenRA.Widgets;
